Moved PageController to it's own subdir
